Last commit before moving to Swift packages
Massive code re-organisation, here are some highlights:
-Renamed `CreationVariablesManager` to `CreationProcess`
-Renamed `CreationVariablesManager.CreationProcess` to `CreationProcess.Management`
-Created new managers for recovery mode and user's name
-Renamed CommandsManager to Command
-Re-organised the options-related stuff
-Made some View's code more simple and optimised
-Optimisations and simplifications for the MediaCreationManager components
- `CreationVariablesManager.OtherOptionsManager` renamed to `CreationProcess.OptionsManager`
- `CreationProcess.OptionsManager.ID` variables renamed with shorter and better name
- Simplified `Part` and centralised some things into it, like the obtainment of the Part's name and icons
- Simplified `DriveView` to work with a generic protocol rather that using booleans
- Improved `ShadowView`
- improved `Command.Sudo` (previously `CommandManager.SudoManager`

Additions:
-Added a `UImanager` to take care of previously-global stuff
-Added `InstallerAppInfo` to manage the installer similarly  to what's done with the disk using the `Part` objects

Fixes:
-Fixed some minimum windows sizes
-Improved some text assets and text readability

Other:
- Removed `simulateDisableShadows` because it was redundant with the current theming system
- Removed a bunch of unnecessary startup actions

NOTE: This commit has not been extensively tested and might have various issues in desperate need of a fix.

+++ TINU/AppBanner.swift
@@ -13,21 +13,21 @@ public final class AppBanner{
 	private static let sideBorder = "\u{2551}"
 	private static let topBorder = "\u{2550}"
 	
-	private static let margin: UInt = 4
+	private static let margin: UInt64 = 4
 	
-	private static var width: UInt{
+	private static var width: UInt64{
 		struct MEM{
-			static var w: UInt = 0
+			static var w: UInt64 = 0
 		}
 		
 		if MEM.w > 0{
 			return MEM.w
 		}
 		
-		var max: UInt = 0
+		var max: UInt64 = 0
 		
 		for c in bannerContents{
-			let cc = UInt(c.count)
+			let cc = UInt64(c.count)
 			if cc > max{
 				max = cc
 			}
@@ -70,7 +70,7 @@ public final class AppBanner{
 		}
 		
 		let empty = getEmptyRow()
-		let vfill = strFill(of: empty, length: margin/2, startSeq: nil, endSeq: nil, forget: true)
+		let vfill = strFill(of: empty, length: margin / 2, startSeq: nil, endSeq: nil, forget: true)
 		let m = strFill(of: " ", length: margin, startSeq: nil, endSeq: nil, forget: true)
 		
 		MEM.cBanner += getRow(true) + vfill
@@ -81,7 +81,7 @@ public final class AppBanner{
 				continue
 			}
 			
-			let len = (width - UInt(c.count))
+			let len = (width - UInt64(c.count))
 			
 			MEM.cBanner += sideBorder + m + c + strFill(of: " ", length: len, startSeq: nil, endSeq: nil, forget: true) + m + sideBorder + "\n"
 		}
